Fix the problem of state not updating
Basically it was because of two things:

- The `copy` builtin function works by copying
from one source to a destination slice. The problem
is that it only copies the min(len(src), len(dst)).
If one of your slices are like empty, nothing will be copied
and that is based on the length and not capacity.

- On the `for select` where I'm listening to channel events,
if you use something like:
	case <-data:
		fmt.Println(<-data)
It is not going to work. I assume it is because it is always
jumping one message. The right way is:
	case x := <-data:
		fmt.Println(x)